title(功能):完成播放器手势模块，实现左侧屏幕亮度，右侧声音改边得ui效果。 同时解决点击事件冲突问题。 reason: study

+++ app/src/main/java/com/kingz/play/gesture/IGestureCallBack.java
@@ -23,7 +23,7 @@ public interface IGestureCallBack {
     /**
      * @param duration :快进快退,大于0快进，小于0快退
      */
-    void onGestureUpdateVideoTime(int duration);
+    void onGestureUpdateVideoTime(long duration);
 
     /**
      * 单击手势，确认是单击的时候调用
